backend_python/requirements.txt
anyio==4.8.0
click==8.1.8
fastapi==0.95.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
pydantic==1.10.21
python-dotenv==1.0.0
python-multipart==0.0.20
setuptools==75.8.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.26.1
typing_extensions==4.12.2
uvicorn==0.22.0
wheel==0.45.1
